+++ src/main/java/com/danila/market/entity/OrderDetails.java
@@ -26,6 +26,6 @@ public class OrderDetails {
     @ManyToOne
     @JoinColumn(name = "product_id")
     private Product product;
-    private BigDecimal amount;
-    private BigDecimal price;
+    private int amount;
+    private double price;
 }
